% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{apimisc_get_balo_siren}
\alias{apimisc_get_balo_siren}
\title{apimisc_get_balo_siren}
\usage{
apimisc_get_balo_siren(siren = "")
}
\arguments{
\item{siren}{le siren a interroger}
}
\value{
un data.frame du resultat de la requete sur siren dans le balo
}
\description{
Retourne le resultat de la requete sur siren dans le balo
}
\examples{
apimisc_get_balo_siren(siren = "123456789")
}
