% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{apimisc_get_jo_cat}
\alias{apimisc_get_jo_cat}
\title{apimisc_get_jo_cat}
\usage{
apimisc_get_jo_cat()
}
\value{
un data.frame des catalogues de l'API JO
}
\description{
Retourne le catalogue de l'API journalofficiel
}
\examples{
\dontrun{
apimisc_get_jo_cat()
}
}
