% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{apimisc_if_get_catalog_list}
\alias{apimisc_if_get_catalog_list}
\title{apimisc_if_get_catalog_list}
\usage{
apimisc_if_get_catalog_list()
}
\value{
un data.frame du catalogue de l'API info fi
}
\description{
Retourne la liste des catalogues de l'API info fi
}
\examples{
apimisc_if_get_catalog_list()
}
