% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{apimisc_if_get_data}
\alias{apimisc_if_get_data}
\title{apimisc_if_get_dataset}
\usage{
apimisc_if_get_data(cat = "", simp = TRUE)
}
\arguments{
\item{cat}{le catalogue demande}
}
\value{
un data.frame du catalogue de l'API info fi
}
\description{
Retourne le catalogue de l'API info fi demande
}
\examples{
apimisc_if_get_dataset(cat = "")
}
