% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{apimisc_make_link}
\alias{apimisc_make_link}
\title{apimisc_make_link}
\usage{
apimisc_make_link(url, text = "clic")
}
\arguments{
\item{url}{l'url a rendre cliquable}

\item{text}{texte a afficher en remplacement de l'url}
}
\value{
une chaine de caracteres cliquable
}
\description{
Retourne la chaine de caractere entouree des balises pour la rendre cliquable
}
\examples{
\dontrun{
apimisc_make_link(url = "monurl", text = "clic")
}
}
