% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{apimisc_make_report}
\alias{apimisc_make_report}
\title{generer un rapport html d'une entreprise}
\usage{
apimisc_make_report(siren)
}
\arguments{
\item{siren}{une chaine de caractere pour l'entreprise a consulter}
}
\value{
un fichier html, affiche dans le browser
}
\description{
Affiche une page html synthétisant de l'information de base
}
\examples{
apimisc_make_report(siren = "401887401")
}
