% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{apimisc_parse_bodacc_siren}
\alias{apimisc_parse_bodacc_siren}
\title{apimisc_parse_bodacc_siren}
\usage{
apimisc_parse_bodacc_siren(x)
}
\arguments{
\item{x}{une requete get_bodacc_siren}
}
\value{
un data.frame du resultat parse de la requete sur siren
}
\description{
Retourne le resultat parse de la requete sur siren dans le bodacc
}
\examples{
apimisc_parse_bodacc_siren(x)
}
