% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{apimisc_template_path}
\alias{apimisc_template_path}
\title{acceder au repertoire du template pour tdb multi api}
\usage{
apimisc_template_path(pack = "apimisc", tempname = "multiapi")
}
\arguments{
\item{pack}{le nom du package}

\item{tempname}{le nom du template}
}
\value{
une chaine de char
}
\description{
Retourne le chemin du template Rmarkdown pour dashboard siren
}
\examples{
\dontrun{
apimisc_template_path(pack = "apimisc")
}
}
