% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{apimisc_www_path}
\alias{apimisc_www_path}
\title{acceder au repertoire www du package}
\usage{
apimisc_www_path(pack = "apimisc")
}
\arguments{
\item{pack}{le nom du package}
}
\value{
une chaine de char
}
\description{
Retourne le chemin du repertoire www
}
\examples{
apimisc_www_path(pack = "apimisc")
}
