% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{bdf_get_catalog}
\alias{bdf_get_catalog}
\title{get banque de france dataset catalog}
\usage{
bdf_get_catalog()
}
\value{
data.frame
}
\description{
retourne un data.frame des datasets du catalogue banque de france
}
\examples{
bdf_get_catalog(dataset = "OFC")
}
