% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{bdf_get_serie}
\alias{bdf_get_serie}
\title{get serie from banque de france dataset catalog}
\usage{
bdf_get_serie(key = NULL)
}
\arguments{
\item{key}{le nom de la serie}
}
\value{
data.frame
}
\description{
retourne un data.frame d'une serie de la banque de france
}
\examples{
\dontrun{
bdf_get_serie(key = "MIR1.M.FR.B.L23FRLA.D.R.A.2230U6.EUR.O")
}
}
