% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{bdf_list_datasets}
\alias{bdf_list_datasets}
\title{get series infos from banque de france dataset catalog}
\usage{
bdf_list_datasets(dataset = "OFC")
}
\arguments{
\item{dataset}{le code du catalogue}
}
\value{
data.frame
}
\description{
retourne un data.frame d'information sur les series d'un dataset du catalogue banque de france
}
\examples{
bdf_list_datasets(dataset = "OFC")
}
