% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datagouv20250920}
\alias{datagouv20250920}
\title{Dataset from data.gouv.fr (2025-09)}
\format{
A data frame with x rows and y variables:
\describe{
\item{access_service}{access_service}
\item{access_url}{access_url}
\item{accrual_periodicity}{accrual_periodicity}
\item{algorithmbyte_size}{algorithmbyte_size}
\item{checksum}{checksum}
\item{checksum_value}{checksum_value}
\item{contact_point}{contact_point}
\item{datasetdct_conforms_to}{datasetdct_conforms_to}
\item{dct_creator}{dct_creator}
\item{description}{description}
\item{distributiondownload_url}{distributiondownload_url}
\item{end_date}{end_date}
\item{endpoint_description}{endpoint_description}
\item{endpoint_url}{endpoint_url}
\item{firstfn}{firstfn}
\item{format}{format}
\item{homepage}{homepage}
\item{http_data_europa_eu_930_distribut}{http_data_europa_eu_930_distribut}
\item{http_data_europa_eu_r5r_applicable_legislation}{http_data_europa_eu_r5r_applicable_legislation}
\item{http_data_europa_eu_r5r_hvd_category}{http_data_europa_eu_r5r_hvd_category}
\item{http_www_w3_org_ns_adms_number_identifier}{http_www_w3_org_ns_adms_number_identifier}
\item{id}{id}
\item{identifier}{identifier}
\item{issued}{issued}
\item{keyword}{keyword}
\item{label}{label}
\item{landing_page}{landing_page}
\item{language}{language}
\item{last}{last}
\item{license}{license}
\item{media_type}{media_type}
\item{modified}{modified}
\item{name}{name}
\item{next.}{next.}
\item{publisher}{publisher}
\item{rights}{rights}
\item{serves_dataset}{serves_dataset}
\item{service}{service}
\item{skos_alt_label}{skos_alt_label}
\item{skos_notation}{skos_notation}
\item{start_date}{start_date}
\item{temporal}{temporal}
\item{title}{title}
\item{total_items}{total_items}
\item{type}{type}
\item{vcard_has_email}{vcard_has_email}
\item{vcard_has_url}{vcard_has_url}
\item{view}{view}
}
}
\usage{
datagouv20250920
}
\description{
Dataset contenant les codes et libelle de la naf niveau 5
Les variables sont les suivantes
}
\keyword{datasets}
