% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{dg_get_site_data_portal}
\alias{dg_get_site_data_portal}
\title{get data.gouv.fr data portal informations}
\usage{
dg_get_site_data_portal()
}
\value{
data.frame
}
\description{
cree un repertoire dans HOME et enregistre un data.frame sur les series de data.gouv.fr
}
\examples{
\dontrun{
dg_get_site_data_portal()
}
}
