% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{dg_read_site_data_portal}
\alias{dg_read_site_data_portal}
\title{read data.gouv.fr data portal informations}
\usage{
dg_read_site_data_portal()
}
\value{
data.frame
}
\description{
lit dans le répertoire HOME le fichier sur les series de data.gouv.fr
}
\examples{
dg_read_site_data_portal()
}
