% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{dg_report_dataservice}
\alias{dg_report_dataservice}
\title{make report on dataservice}
\usage{
dg_report_dataservice(ident)
}
\arguments{
\item{ident}{l'identifiant du dataservice}
}
\value{
flexdashboard report
}
\description{
cree un rapport sur un dataservice
}
\examples{
\dontrun{
dg_report_dataservice(ident = NULL)
}
}
