% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtprod}
\alias{dtprod}
\title{dtprod}
\format{
A data frame with x rows and y variables:
\describe{
\item{api}{nom de l'api}
\item{cat}{categorie de dataset}
\item{lien}{url de l'api}
\item{package}{package}
\item{package_url}{url du package}
\item{portail}{nom du portail}
\item{prod}{producteur}
\item{restriction}{restriction d'usage}
\item{service}{service}
\item{theme}{theme}
}
}
\usage{
dtprod
}
\description{
Dataset contenant ...
Les variables sont les suivantes
}
\keyword{datasets}
