% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{if_get_catalog_list}
\alias{if_get_catalog_list}
\title{get info-fi API catalog}
\usage{
if_get_catalog_list()
}
\value{
un data.frame du catalogue de l'API info fi
}
\description{
Retourne la liste des catalogues de l'API info fi
}
\examples{
\dontrun{
if_get_catalog_list()
}
}
