% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{if_get_dataset}
\alias{if_get_dataset}
\title{get info-fi API dataset}
\usage{
if_get_dataset(cat = "", simp = TRUE)
}
\arguments{
\item{cat}{le catalogue demande}

\item{simp}{if TRUE (default) simplify vector}
}
\value{
un data.frame du catalogue de l'API info fi
}
\description{
Retourne le catalogue de l'API info fi demande
}
\examples{
\dontrun{
if_get_dataset(cat = "")
}
}
