% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{imf_get_country}
\alias{imf_get_country}
\title{get countries code}
\usage{
imf_get_country()
}
\value{
data.frame
}
\description{
telecharge les codes pays et lib en fr (iso 3166 alpha 3)
}
\examples{
imf_get_country()
}
