% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{imf_get_dataset_info}
\alias{imf_get_dataset_info}
\title{get dataset informations}
\usage{
imf_get_dataset_info()
}
\value{
data.frame
}
\description{
telecharge des infos sur les datasets dispo sur IMF
}
\examples{
\dontrun{
imf_get_dataset_info()
}
}
