% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{oecd_get_catalog}
\alias{oecd_get_catalog}
\title{get oecd catalog}
\usage{
oecd_get_catalog()
}
\value{
data.frame
}
\description{
telecharge les catalogues de donnees de l'ocde
}
\examples{
\dontrun{
oecd_get_catalog()
}
}
