% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soccac40}
\alias{soccac40}
\title{Dataset contenant les info d'info-fin
Les variables sont les suivantes}
\format{
A data frame with x rows and y variables:
\describe{
\item{_id}{identifiant}
\item{_links}{liens}
\item{_size}{taille}
\item{_timestamp}{horodatage}
\item{cac40}{societe du CAC 40}
\item{first_letter}{premiere lettre}
\item{id}{ident2}
\item{isin}{isin}
\item{name}{nom}
\item{name_cac40}{nom au CAC 40}
\item{rcs}{rcs}
}
}
\usage{
soccac40
}
\description{
Dataset contenant les info d'info-fin
Les variables sont les suivantes
}
\keyword{datasets}
