% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{wb_get_catalog}
\alias{wb_get_catalog}
\title{get world bank catalog}
\usage{
wb_get_catalog()
}
\value{
data.frame
}
\description{
telecharge des infos sur les datasets dispo via World Bank
}
\examples{
\dontrun{
wb_get_catalog()
}
}
