% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apimiscutils.R
\name{apimisc_parse_jo_cat}
\alias{apimisc_parse_jo_cat}
\title{apimisc_parse_jo_cat}
\usage{
apimisc_parse_jo_cat(cat, limite = TRUE)
}
\arguments{
\item{cat}{le catalogue a parser}

\item{limite}{si TRUE (par defaut), limite a 4 le nb de variables renvoyees par la fonction}
}
\value{
un data.frame des catalogues de l'API JO parse
}
\description{
Retourne le catalogue de l'API journalofficiel parse
}
\examples{
\dontrun{apimisc_parse_jo_cat(cat = "moncat")}
}
