% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datagouv20250920}
\alias{datagouv20250920}
\title{Dataset from data.gouv.fr (2025-09)}
\format{
A data frame with x rows and y variables:
\describe{
\item{distribution}{distribution}
\item{id}{id}
\item{keyword}{keyword}
\item{title}{title}
\item{type}{type}
}
}
\usage{
datagouv20250920
}
\description{
Dataset contenant des infos sur les donnees dispo sur data.gouv
Les variables sont les suivantes
}
\keyword{datasets}
